Refactor imports in gwkokab/errors.py and fix variable names in banana_error_m1_m2 function
