+++ src/kokab/single_gaussians/genie.py
@@ -24,7 +24,7 @@ from numpyro.distributions import MultivariateNormal
 from gwkokab.errors import banana_error_m1_m2
 from gwkokab.parameters import PRIMARY_MASS_SOURCE, SECONDARY_MASS_SOURCE
 from gwkokab.population import error_magazine, popfactory, popmodel_magazine
-from gwkokab.vts._neuralvt import load_model
+from gwkokab.vts import load_model
 
 from ..utils import genie_parser
